[fix](be) Fix row source buffer overflow in vertical compaction (#83)
* [fix](be) Fix row source buffer overflow in vertical compaction

### What problem does this PR solve?

Issue Number: N/A

Related PR: apache/doris#63152, apache/doris#53972

Problem Summary: Fix RowSourcesBuffer spill capacity accounting and 32-bit traversal narrowing that can crash vertical compaction for very large tablets.

### Release note

Fix a BE crash during vertical compaction of very large tablets.

### Check List (For Author)

- Test: Not run as requested; added a RowSourcesBuffer spill regression test
- Behavior changed: Yes, RowSourcesBuffer spills before an append would exceed usable capacity
- Does this need documentation: No

* [fix](be) Fix signed overflow in row source buffer memory threshold

* [fix](be) Fix row source batch start after spill

### What problem does this PR solve?

Issue Number: N/A

Related PR: #83

Problem Summary: UNIQUE vertical compaction could reuse the previous row-source batch start when a spill produced the same buffered size, causing delete flags to be applied to stale positions.

### Release note

Fix UNIQUE-key vertical compaction failures at equal-size row-source spill boundaries.

### Check List (For Author)

- Test: Not run (per request; added BE unit-test coverage)
- Behavior changed: Yes (derive the current row-source batch start from the total-size increment)
- Does this need documentation: No

+++ be/src/vec/olap/vertical_merge_iterator.h
@@ -104,7 +104,7 @@ public:
         DCHECK(_buf_idx < _buffer->size());
         return RowSource(_buffer->get_element(_buf_idx));
     }
-    void advance(int32_t step = 1) {
+    void advance(int64_t step = 1) {
         DCHECK(_buf_idx + step <= _buffer->size());
         _buf_idx += step;
     }
